The HTML now receivces some data, the java endpoint is ready

+++ first_spring_project/src/main/resources/static/js/users.js
@@ -37,8 +37,9 @@ function fillTable(users) {
 }
 
 function removeUser(id) {
+  if (!confirm("Are you user you want to remove the user?")) return;
   fetch(`${API_BASE_URL}/users/${id}`, {
     method: "DELETE",
     headers,
-  })
+  });
 }
